requirements.txt
asgiref==3.8.1
build==1.2.2.post1
click==8.1.8
colorama==0.4.6
coverage==7.6.4
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
packaging==24.2
pip-tools==7.4.1
psycopg2==2.9.10
PyJWT==2.10.1
pyproject_hooks==1.2.0
python-decouple==3.8
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
wheel==0.45.1
